requirements.txt
anthropic>=0.92.0
colorama>=0.4.6
matplotlib>=3.7.0
python-dotenv>=1.0.0
Pillow>=10.0.0
